requirements.txt
Flask==2.2.5
pymongo==4.7.0
python-dotenv==1.0.0
